chore: test case issue resolved
